+++ test/integration/role-azure-flow.test.ts
@@ -10,13 +10,13 @@ import { MockArmState, createMockArmServer } from "../mock-arm.js";
 import { MockGraphState, createMockGraphServer } from "../mock-graph.js";
 import { fetchCsrfToken, testSignal } from "../helpers.js";
 
-import { pimRoleAzureActiveListTool } from "../../src/tools/pim/role-azure/pim-role-azure-active-list.js";
-import { pimRoleAzureApprovalListTool } from "../../src/tools/pim/role-azure/pim-role-azure-approval-list.js";
-import { pimRoleAzureApprovalReviewTool } from "../../src/tools/pim/role-azure/pim-role-azure-approval-review.js";
-import { pimRoleAzureDeactivateTool } from "../../src/tools/pim/role-azure/pim-role-azure-deactivate.js";
-import { pimRoleAzureEligibleListTool } from "../../src/tools/pim/role-azure/pim-role-azure-eligible-list.js";
-import { pimRoleAzureRequestListTool } from "../../src/tools/pim/role-azure/pim-role-azure-request-list.js";
-import { pimRoleAzureRequestTool } from "../../src/tools/pim/role-azure/pim-role-azure-request.js";
+import { pimRoleAzureActiveListTool } from "../../src/features/role-azure/tools/pim-role-azure-active-list.js";
+import { pimRoleAzureApprovalListTool } from "../../src/features/role-azure/tools/pim-role-azure-approval-list.js";
+import { pimRoleAzureApprovalReviewTool } from "../../src/features/role-azure/tools/pim-role-azure-approval-review.js";
+import { pimRoleAzureDeactivateTool } from "../../src/features/role-azure/tools/pim-role-azure-deactivate.js";
+import { pimRoleAzureEligibleListTool } from "../../src/features/role-azure/tools/pim-role-azure-eligible-list.js";
+import { pimRoleAzureRequestListTool } from "../../src/features/role-azure/tools/pim-role-azure-request-list.js";
+import { pimRoleAzureRequestTool } from "../../src/features/role-azure/tools/pim-role-azure-request.js";
 
 interface Harness {
   config: ServerConfig;