+++ test/tools/pim/role-entra/list-tools.test.ts
@@ -9,10 +9,10 @@ import type { ServerConfig } from "../../../../src/index.js";
 import { MockGraphState, createMockGraphServer } from "../../../mock-graph.js";
 import { testSignal } from "../../../helpers.js";
 
-import { pimRoleEntraActiveListTool } from "../../../../src/tools/pim/role-entra/pim-role-entra-active-list.js";
-import { pimRoleEntraApprovalListTool } from "../../../../src/tools/pim/role-entra/pim-role-entra-approval-list.js";
-import { pimRoleEntraEligibleListTool } from "../../../../src/tools/pim/role-entra/pim-role-entra-eligible-list.js";
-import { pimRoleEntraRequestListTool } from "../../../../src/tools/pim/role-entra/pim-role-entra-request-list.js";
+import { pimRoleEntraActiveListTool } from "../../../../src/features/role-entra/tools/pim-role-entra-active-list.js";
+import { pimRoleEntraApprovalListTool } from "../../../../src/features/role-entra/tools/pim-role-entra-approval-list.js";
+import { pimRoleEntraEligibleListTool } from "../../../../src/features/role-entra/tools/pim-role-entra-eligible-list.js";
+import { pimRoleEntraRequestListTool } from "../../../../src/features/role-entra/tools/pim-role-entra-request-list.js";
 
 interface ToolResult {
   content: { type: string; text: string }[];